Conflicts:
	core/lib/hadoop-0.20.2/CHANGES.txt
	core/lib/hadoop-0.20.2/LICENSE.txt
	core/lib/hadoop-0.20.2/NOTICE.txt
	core/lib/hadoop-0.20.2/README.txt
	core/lib/hadoop-0.20.2/bin/hadoop
	core/lib/hadoop-0.20.2/bin/hadoop-config.sh
	core/lib/hadoop-0.20.2/bin/hadoop-daemon.sh
	core/lib/hadoop-0.20.2/bin/hadoop-daemons.sh
	core/lib/hadoop-0.20.2/bin/rcc
	core/lib/hadoop-0.20.2/bin/slaves.sh
	core/lib/hadoop-0.20.2/bin/start-all.sh
	core/lib/hadoop-0.20.2/bin/start-balancer.sh
	core/lib/hadoop-0.20.2/bin/start-dfs.sh
	core/lib/hadoop-0.20.2/bin/start-mapred.sh
	core/lib/hadoop-0.20.2/bin/stop-all.sh
	core/lib/hadoop-0.20.2/bin/stop-balancer.sh
	core/lib/hadoop-0.20.2/bin/stop-dfs.sh
	core/lib/hadoop-0.20.2/bin/stop-mapred.sh
	core/lib/hadoop-0.20.2/build.xml
	core/lib/hadoop-0.20.2/conf/configuration.xsl
	core/lib/hadoop-0.20.2/conf/core-site.xml
	core/lib/hadoop-0.20.2/conf/hadoop-env.sh
	core/lib/hadoop-0.20.2/conf/hadoop-metrics.properties
	core/lib/hadoop-0.20.2/conf/hadoop-policy.xml
	core/lib/hadoop-0.20.2/conf/hdfs-site.xml
	core/lib/hadoop-0.20.2/conf/log4j.properties
	core/lib/hadoop-0.20.2/conf/mapred-site.xml
	core/lib/hadoop-0.20.2/conf/masters
	core/lib/hadoop-0.20.2/conf/slaves
	core/lib/hadoop-0.20.2/conf/ssl-client.xml.example
	core/lib/hadoop-0.20.2/conf/ssl-server.xml.example
	core/lib/hadoop-0.20.2/hadoop-0.20.2-ant.jar
	core/lib/hadoop-0.20.2/hadoop-0.20.2-core.jar
	core/lib/hadoop-0.20.2/hadoop-0.20.2-examples.jar
	core/lib/hadoop-0.20.2/hadoop-0.20.2-test.jar
	core/lib/hadoop-0.20.2/hadoop-0.20.2-tools.jar
	core/lib/hadoop-0.20.2/ivy.xml
	core/lib/hadoop-0.20.2/ivy/hadoop-core.pom
	core/lib/hadoop-0.20.2/ivy/ivy-2.0.0-rc2.jar
	core/lib/hadoop-0.20.2/ivy/ivysettings.xml
	core/lib/hadoop-0.20.2/ivy/libraries.properties
	core/lib/hadoop-0.20.2/lib/commons-codec-1.3.jar
	core/lib/hadoop-0.20.2/lib/commons-el-1.0.jar
	core/lib/hadoop-0.20.2/lib/commons-httpclient-3.0.1.jar
	core/lib/hadoop-0.20.2/lib/commons-logging-1.0.4.jar
	core/lib/hadoop-0.20.2/lib/commons-logging-api-1.0.4.jar
	core/lib/hadoop-0.20.2/lib/commons-net-1.4.1.jar
	core/lib/hadoop-0.20.2/lib/core-3.1.1.jar
	core/lib/hadoop-0.20.2/lib/hsqldb-1.8.0.10.LICENSE.txt
	core/lib/hadoop-0.20.2/lib/hsqldb-1.8.0.10.jar
	core/lib/hadoop-0.20.2/lib/jasper-compiler-5.5.12.jar
	core/lib/hadoop-0.20.2/lib/jasper-runtime-5.5.12.jar
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.17.0.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.18.1.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.18.2.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.18.3.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.19.0.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.19.2.xml
	core/lib/hadoop-0.20.2/lib/jdiff/hadoop_0.20.2.xml
	core/lib/hadoop-0.20.2/lib/jets3t-0.6.1.jar
	core/lib/hadoop-0.20.2/lib/jetty-6.1.14.jar
	core/lib/hadoop-0.20.2/lib/jetty-util-6.1.14.jar
	core/lib/hadoop-0.20.2/lib/jsp-2.1/jsp-2.1.jar
	core/lib/hadoop-0.20.2/lib/jsp-2.1/jsp-api-2.1.jar
	core/lib/hadoop-0.20.2/lib/junit-3.8.1.jar
	core/lib/hadoop-0.20.2/lib/kfs-0.2.2.jar
	core/lib/hadoop-0.20.2/lib/kfs-0.2.LICENSE.txt
	core/lib/hadoop-0.20.2/lib/log4j-1.2.15.jar
	core/lib/hadoop-0.20.2/lib/native/Linux-amd64-64/libhadoop.la
	core/lib/hadoop-0.20.2/lib/native/Linux-i386-32/libhadoop.la
	core/lib/hadoop-0.20.2/lib/oro-2.0.8.jar
	core/lib/hadoop-0.20.2/lib/servlet-api-2.5-6.1.14.jar
	core/lib/hadoop-0.20.2/lib/xmlenc-0.52.jar
	core/src/main/scala/spark/LocalScheduler.scala
	project/build/SparkProject.scala
	run

+++ core/src/main/scala/spark/MesosScheduler.scala
@@ -21,7 +21,7 @@ import mesos._
  */
 private class MesosScheduler(
   sc: SparkContext, master: String, frameworkName: String)
-extends MScheduler with spark.Scheduler with Logging
+extends MScheduler with DAGScheduler with Logging
 {
   // Environment variables to pass to our executors
   val ENV_VARS_TO_SEND_TO_EXECUTORS = Array(
@@ -52,7 +52,7 @@ extends MScheduler with spark.Scheduler with Logging
 
   // URIs of JARs to pass to executor
   var jarUris: String = ""
-
+  
   def newJobId(): Int = this.synchronized {
     val id = nextJobId
     nextJobId += 1
@@ -101,29 +101,27 @@ extends MScheduler with spark.Scheduler with Logging
     new ExecutorInfo(execScript, createExecArg(), params)
   }
 
-  /**
-   * The primary means to submit a job to the scheduler. Given a list of tasks,
-   * runs them and returns an array of the results.
-   */
-  override def runTasks[T: ClassManifest](tasks: Array[Task[T]]): Array[T] = {
+  
+  def submitTasks(tasks: Seq[Task[_]]) {
+    logInfo("Got a job with " + tasks.size + " tasks")
     waitForRegister()
-    val jobId = newJobId()
-    val myJob = new SimpleJob(this, tasks, jobId)
-    try {
-      this.synchronized {
-        activeJobs(jobId) = myJob
-        activeJobsQueue += myJob
-        jobTasks(jobId) = new HashSet()
-      }
-      driver.reviveOffers();
-      return myJob.join();
-    } finally {
-      this.synchronized {
-        activeJobs -= jobId
-        activeJobsQueue.dequeueAll(x => (x == myJob))
-        taskIdToJobId --= jobTasks(jobId)
-        jobTasks.remove(jobId)
-      }
+    this.synchronized {
+      val jobId = newJobId()
+      val myJob = new SimpleJob(this, tasks, jobId)
+      activeJobs(jobId) = myJob
+      activeJobsQueue += myJob
+      logInfo("Adding job with ID " + jobId)
+      jobTasks(jobId) = new HashSet()
+    }
+    driver.reviveOffers();
+  }
+  
+  def jobFinished(job: Job) {
+    this.synchronized {
+      activeJobs -= job.getId
+      activeJobsQueue.dequeueAll(x => (x == job))
+      taskIdToJobId --= jobTasks(job.getId)
+      jobTasks.remove(job.getId)
     }
   }
 
@@ -199,10 +197,12 @@ extends MScheduler with spark.Scheduler with Logging
             }
             if (isFinished(status.getState)) {
               taskIdToJobId.remove(status.getTaskId)
-              jobTasks(jobId) -= status.getTaskId
+              if (jobTasks.contains(jobId))
+                jobTasks(jobId) -= status.getTaskId
             }
           case None =>
-            logInfo("TID " + status.getTaskId + " already finished")
+            logInfo("Ignoring update from TID " + status.getTaskId +
+              " because its job is gone")
         }
       } catch {
         case e: Exception => logError("Exception in statusUpdate", e)
@@ -253,9 +253,11 @@ extends MScheduler with spark.Scheduler with Logging
     // Copy each JAR to a unique filename in the jarDir
     for ((path, index) <- sc.jars.zipWithIndex) {
       val file = new File(path)
-      val filename = index + "_" + file.getName
-      copyFile(file, new File(jarDir, filename))
-      filenames += filename
+      if (file.exists) {
+        val filename = index + "_" + file.getName
+        copyFile(file, new File(jarDir, filename))
+        filenames += filename
+      }
     }
     // Create the server
     jarServer = new HttpServer(jarDir)
